Refactor: 페이지 - 템플릿 관심사 분리
